% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AvGPR.R
\name{AvGPR}
\alias{AvGPR}
\title{Calculates an average Gaussian Process}
\usage{
AvGPR(design, response, test, nWeight = 3)
}
\arguments{
\item{design}{Design data frame}

\item{response}{Response data frame}

\item{test}{Test data frame}

\item{nWeight}{Number of weight calculation simulations}
}
\value{
Object of R3 class average Gaussian Process
}
\description{
Calculates an average Gaussian Process
}
\examples{
n <- 100
X <- seq(0, 4 * pi, length.out = n)
Y <- 2 * sin(X)
XX <- seq(1,10, length.out=15)
AvGPR(data.frame(X), data.frame(Y), data.frame(X=XX))

}
