% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AvGPR_GF.R
\name{AvGPR_GF}
\alias{AvGPR_GF}
\title{Calculates the weighting statistic}
\usage{
AvGPR_GF(y_pred, y_true, var, sd = FALSE)
}
\arguments{
\item{y_pred}{vector of mean predicted values}

\item{y_true}{vector of true values}

\item{var}{vector of variances}

\item{bool}{if 'var' is standard deviation}
}
\value{
weighting statistic
}
\description{
Calculates the weighting statistic
}
\examples{
y_pred <- c(1, 2, 4.1, 5.2)
y_true <- c(1.5, 2.5, 3.5, 4.5)
var <- c(0.6, 0.8, 1, 1.1)
AvGPR_GF(y_pred, y_true, var, sd = TRUE)

}
