% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AvGPR_weight.R
\name{AvGPR_weight}
\alias{AvGPR_weight}
\title{Calculates the model weights}
\usage{
AvGPR_weight(values)
}
\arguments{
\item{values}{vector of the weighting statistic}
}
\value{
Model weights
}
\description{
Calculates the model weights
}
\examples{
AvGPR_weight(c(12,24,21,16,11))

}
